Implement wrapping GC for Linux #61
This renames the original native GC class to NativeGC and introduced a
GC wrapper class that serves as a delegate for an inner GC
implementation, which may either be the existing NativeGC or a SkijaGC
(or any other GCHandle implemented in the future). It aligns the
implementation with the changes already made for the Windows
implementation of SWT.

Contributes to
https://github.com/swt-initiative31/prototype-skija/issues/61
