+++ bundles/org.eclipse.swt/Eclipse SWT/cairo/org/eclipse/swt/graphics/Path.java
@@ -337,7 +337,7 @@ public void addString(String string, float x, float y, Font font) {
 }
 void addStringInPixels(String string, float x, float y, Font font) {
 	moved = false;
-	GC.addCairoString(handle, string, x, y, font);
+	NativeGC.addCairoString(handle, string, x, y, font);
 	closed = true;
 }
 
@@ -393,7 +393,7 @@ boolean containsInPixels(float x, float y, GC gc, boolean outline) {
 	gc.initCairo();
 	gc.checkGC(GC.LINE_CAP | GC.LINE_JOIN | GC.LINE_STYLE | GC.LINE_WIDTH);
 	boolean result = false;
-	long cairo = gc.data.cairo;
+	long cairo = gc.getGCData().cairo;
 	long copy = Cairo.cairo_copy_path(handle);
 	if (copy == 0) SWT.error(SWT.ERROR_NO_HANDLES);
 	Cairo.cairo_append_path(cairo, copy);
