+++ bundles/org.eclipse.swt/Eclipse SWT/gtk/org/eclipse/swt/widgets/Control.java
@@ -269,7 +269,8 @@ boolean drawGripper (GC gc, int x, int y, int width, int height, boolean vertica
 	GTK.gtk_style_context_save (context);
 	GTK.gtk_style_context_add_class (context, GTK.GTK_STYLE_CLASS_PANE_SEPARATOR);
 	GTK.gtk_style_context_set_state (context, GTK.GTK_STATE_FLAG_NORMAL);
-	GTK.gtk_render_handle (context, gc.handle, x, y, width, height);
+	NativeGC ngc = (NativeGC) gc.innerGC;
+	GTK.gtk_render_handle(context, ngc.handle, x, y, width, height);
 	GTK.gtk_style_context_restore (context);
 	return true;
 }
@@ -652,7 +653,8 @@ public boolean print (GC gc) {
 	// Prevent allocation warnings
 	GTK.gtk_widget_get_preferred_size(topHandle, null, null);
 	GTK3.gtk_widget_size_allocate(topHandle, allocation);
-	GTK3.gtk_widget_draw(topHandle, gc.handle);
+	NativeGC ngc = (NativeGC) gc.innerGC;
+	GTK3.gtk_widget_draw(topHandle, ngc.handle);
 	return true;
 }
